[python] Move LocalMemoryObject into the runtime test package
With the local execution path removed, LocalMemoryObject is used only as
an in-memory test fixture. Relocate it from flink_agents.runtime to
flink_agents.runtime.tests so it no longer sits in the production package.
