+++ python/flink_agents/runtime/tests/test_memory_reference.py
@@ -17,7 +17,7 @@
 #################################################################################
 from flink_agents.api.memory_object import MemoryType
 from flink_agents.api.memory_reference import MemoryRef
-from flink_agents.runtime.local_memory_object import LocalMemoryObject
+from flink_agents.runtime.tests.local_memory_object import LocalMemoryObject
 
 
 class MockRunnerContext:
